requirements.txt
flask
tensorflow
pillow